Rename maps_services file
